Move Ducaheat planner into planner package
