+++ custom_components/termoweb/planner/__init__.py
@@ -1,0 +1,1 @@
+"""Planner package for vendor-specific sequencing."""
